+++ collective/trustedimports/libicalendar.py
@@ -1,5 +1,6 @@
 from collective.trustedimports.util import whitelist_module
 
+whitelist_module("pytz")
 whitelist_module(
     "icalendar",
     classes=[